pyproject.toml
[tool.poetry]
name = "aphasia"
version = "0.1.0"
description = "Aphasia type and severity detection from MRI scans experiment toolkit"
authors = ["Matvey Kairov <mkairov@hse.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9.6, <3.12"
jupyter = "^1.0.0"
ipykernel = "^6.29.3"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
python-dotenv = "^1.0.1"

numpy = "^1.24.3"
pandas = "^2.2.0"
openpyxl = "^3.1.2"
nibabel = "^5.2.1"
matplotlib = "^3.8.0"
seaborn = "^0.13.2"
pillow = "^10.2.0"
wandb = "^0.16.3"
tqdm = "^4.66.1"

tensorflow = [
    {version = "^2.16.1", extras = ["and-cuda"], markers = "sys_platform != 'darwin'"},
    {version = "^2.16.1", markers = "sys_platform == 'darwin'"}
]
tensorflow-metal = {version = "^1.1.0", platform = "darwin"}
tensorflow-io-gcs-filesystem = "^0.34.0"
keras = "^3.0.0"

scikit-learn = "^1.2.2"
scikit-optimize = "^0.9.0"
scipy = "^1.12.0"
catboost = "^1.2.2"
xgboost = "^2.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
